Includes: Image classifier, lot detector and utils
This includes the image classifier that trains model.p. I could not find this code in Felipe's repositories but it is a commented replica from one of his video examples.

The file directories are specialised for the way I saved the files, therefore that should be changed to run on your local machine. Make sure to download all the data from this repository, including all masks, videos and ML training data.